requirements.txt
Flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
requests==2.31.0
boto3==1.28.25
Pillow==10.0.0
paho-mqtt==1.6.4